ats-app/requirements.txt
Flask==3.1.0
flask_cors==5.0.1
Flask_Login==0.6.3
mysql_connector_repackaged==0.3.1
python_dateutil==2.9.0.post0
Werkzeug==3.1.3
